ENAHO\CORTE TRANSVERSAL\2004-2021
National Household Survey (ENAHO) published by INEI - Peru.
Link: https://proyectos.inei.gob.pe/microdatos/ 
The objective of the ENAHO is to generate indicators to know the evolution of poverty, welfare and living conditions of households, to make diagnoses on the living conditions and poverty of the population.
LEVEL OF INFERENCE
The survey sample allows us to obtain estimates of the sociodemographic characteristics of the population for the levels of inference: 
From the Integrated sample (panel and non-panel). 
Annual
- National
- Urban National
- Rural National
- 24 departments, each as study domain
- Urban Coast
- Rural Coast
- Urban Highlands
- Rural Highlands
- Urban Jungle
- Rural Jungle
- Metropolitan Area of Lima and Callao
Quarterly
- National
- Urban National
- Rural National

In this case, only non-panel data is used, on an annual basis.

Note that the INEI-ENAHO data sheet and dictionary should be reviewed.